api/pyproject.toml
[tool.poetry]
name = "cs-platform"
version = "0.1.0"
description = ""
authors = ["Zach Pang <zachery.pang@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9.6"
Django = "^3.2.6"
psycopg2 = "^2.9.1"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.8.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
pytest-django = "^4.4.0"
factory-boy = "^3.2.0"
pytest-env = "^0.6.2"

[tool.isort]
# make it compatible with black
profile = "black" 
sections = ['FUTURE', 'STDLIB', 'DJANGO', 'THIRDPARTY','FIRSTPARTY', 'LOCALFOLDER']
known_django = ['django', 'rest_framework']

[tool.pytest.ini_options]
python_files = "test*.py"
DJANGO_SETTINGS_MODULE = "config.settings"
addopts = "--reuse-db"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
